+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## unreleased
 
+- feat: added --prettify flag for re-formatting existing tables
 - feat: added -v/--version flag and colorized clap help output
 - feat: implemented tablefmt CLI